Refactoring: new "snapshots" module
