+++ crates/worker/src/partition/snapshots/mod.rs
@@ -1,0 +1,2 @@
+pub mod producer;
+pub mod repository;
